Simplify length check
By activating the break when the length is *greater than the limit*, not
greater than *or equal*, it is no longer necessary to check the next
char to see if it's
- a space that can be broken on
- the end of the string

Because of this, these inefficient checks can be removed.
As a result of this, though, wrapping in the middle of the word would
add an additional char to the line if the word was wrapped at the index.
To fix this, the word is now wrapped at the previous index.

Resolves #5